Time: 9 ms (18.36%), Space: 57.7 MB (80.69%) - LeetHub

+++ 1631-number-of-sub-arrays-with-odd-sum/1631-number-of-sub-arrays-with-odd-sum.java
@@ -1,6 +1,6 @@
 class Solution {
     public int numOfSubarrays(int[] arr) {
-        int MOD=1000000007;
+        long MOD=1000000007;
         long es=1;
         long os=0;
         long res=0;
